% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuditFairness.R
\name{Audit_Fairness}
\alias{Audit_Fairness}
\title{Wrapper function for fairness auditing.}
\usage{
Audit_Fairness(
  Y,
  S,
  A,
  threshold = 0.5,
  method = "Infairness",
  W = NULL,
  basis = "Poly(S)",
  ...
)
}
\arguments{
\item{Y}{Outcome variable, may contain missing values.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Threshold for classification based on the model score.
Default is 0.5.}

\item{method}{Fairness estimation method. Options are "supervised" or
"semi-supervised". Default is "semi-supervised".}

\item{W}{Optional covariates matrix for adjustment in the semi-supervised
setting. Default is NULL.}

\item{basis}{Basis expansion method for score augmentation. Options include
"Poly(S)", "Poly(S, W)"}

\item{...}{Additional parameters passed to `SSFairness()`.}
}
\value{
A list containing the fairness audit results.
}
\description{
Wrapper function for fairness auditing.
}
