% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Data generation.}
\usage{
DataGeneration(n_labeled, N_unlabeled, prot_att_prevalence, model, rho = 0.4)
}
\arguments{
\item{n_labeled}{Number of labeled examples.}

\item{N_unlabeled}{Number of unlabeled examples.}

\item{prot_att_prevalence}{Prevalence of the protected attribute. For a
single value, this is interpreted as `P(A = 1)`. For a length-two vector,
entries are interpreted as the relative prevalences of groups 0 and 1.}

\item{model}{Indicator to generate from which model.
possible options: "scenario 1", "scenario 2"}

\item{rho}{Correlation between covariates. Default is 0.4.}
}
\value{
Data.frame.
}
\description{
Data generation.
}
