% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisSelection.R
\name{find_alpha_glm}
\alias{find_alpha_glm}
\title{Polynomial order selection via generalized BIC.}
\usage{
find_alpha_glm(
  Y,
  covariates_matrix,
  additional_matrix = NULL,
  gamma = 1:10,
  weights = NULL
)
}
\arguments{
\item{Y}{Binary response variable.}

\item{covariates_matrix}{Covariates matrix.}

\item{additional_matrix}{Additional matrix.}

\item{gamma}{Candidate polynomial orders.}

\item{weights}{Optional perturbation weights.}
}
\value{
Optimal polynomial order.
}
\description{
Polynomial order selection via generalized BIC.
}
