% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{FitParametricCalibration}
\alias{FitParametricCalibration}
\title{Parametric calibration models.}
\usage{
FitParametricCalibration(
  Y_labeled,
  S_labeled,
  A_labeled,
  A_val,
  S_unlabeled,
  method = "Platt",
  W = NULL
)
}
\arguments{
\item{Y_labeled}{Outcome in labeled dataset.}

\item{S_labeled}{Model score in labeled dataset.}

\item{A_labeled}{Group indicator in labeled dataset.}

\item{S_unlabeled}{Model score in unlabeled dataset.}

\item{method}{Method to use; Platt scaling or Beta calibration.
Default is Platt scaling.}
}
\description{
Parametric calibration models.
}
