% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{get_metric}
\alias{get_metric}
\title{Compute fairness metrics by group.}
\usage{
get_metric(Y, S, A, threshold = 0.5, W = NULL)
}
\arguments{
\item{Y}{Outcome or imputed outcome.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Classification threshold. Can be a scalar or a vector
with one value per observation.}

\item{W}{Optional observation weights.}
}
\description{
Compute fairness metrics by group.
}
