% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infairness.R
\name{Infairness}
\alias{Infairness}
\title{Semi-supervised fairness estimation.}
\usage{
Infairness(
  Y,
  S,
  A,
  threshold = 0.5,
  nknots = 3,
  penalty = 1.5,
  W = NULL,
  method = "spline",
  ridge = TRUE,
  step = 1
)
}
\arguments{
\item{Y}{Outcome, can contain missing values.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Threshold for classification based on the model score.
Default value is 0.5.}

\item{nknots}{Number of knots for basis expansion.
Default value is 3.}

\item{W}{Weight for labeled data.
Default value is NULL.}

\item{method}{Smoothing method. Options are  natural cubic spline "spline" , kernel smoothing "ks" , and quadratic polynomial "quad"}

\item{ridge}{Whether to perform ridge regression}

\item{labeled_ind}{Index for labeled data.}

\item{augment_score}{Indicator to apply basis expansion to the score.
Default value is TRUE.}

\item{lambda}{Ridge penalty parameter vector.}
}
\description{
Semi-supervised fairness estimation.
}
