% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infairness_nobasis_temp.R
\name{Infairness_nobasis}
\alias{Infairness_nobasis}
\title{Semi-supervised fairness estimation.}
\usage{
Infairness_nobasis(Y, S, A, threshold = 0.5)
}
\arguments{
\item{Y}{Outcome, can contain missing values.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Threshold for classification based on the model score.}
}
\description{
Semi-supervised fairness estimation.
}
