% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{resample}
\alias{resample}
\title{Purpose: Perturbation Resampling
Updated: 2022-08-30}
\usage{
resample(Y, S, A, method, threshold = 0.5, n_boot = 500, ...)
}
\arguments{
\item{Y}{Outcome in labeled dataset.}

\item{S}{Model score in labeled dataset.}

\item{A}{Group indicator in labeled dataset.}

\item{method}{Group fairness estimation method. Choose from "supervised" or "spline" or "ks".}

\item{threshold}{Threshold for classification based on the model score.
Default value is 0.5.}
}
\description{
Purpose: Perturbation Resampling
Updated: 2022-08-30
}
