% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.R
\name{sum.I}
\alias{sum.I}
\title{Efficiently count comparison relations based on ranks.}
\usage{
\method{sum}{I}(yy, FUN, Yi, Vi = NULL, ties.method = "first")
}
\arguments{
\item{yy}{Vector of values to compare.}

\item{FUN}{Comparison operator such as `"<"` or `"<="`.}

\item{Yi}{Reference vector.}

\item{Vi}{Optional matrix of values to cumulatively aggregate.}

\item{ties.method}{Tie handling method passed to `rank()`.}
}
\description{
Efficiently count comparison relations based on ranks.
}
