% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegacySplineUtils.R
\name{trunc.cub}
\alias{trunc.cub}
\title{Truncated cubic basis helper.}
\usage{
\method{trunc}{cub}(X, x)
}
\arguments{
\item{X}{Variable of interest.}

\item{x}{Knot location.}
}
\description{
Truncated cubic basis helper.
}
