% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyBasis.R
\name{compute_gbic}
\alias{compute_gbic}
\title{Polynomial basis with order minimizing generalized BIC}
\usage{
compute_gbic(
  Y,
  covariates_matrix,
  weights = NULL,
  P_total,
  gamma_ebic = 0.5,
  penalize_intercept = FALSE
)
}
\arguments{
\item{Y}{binary response variable}

\item{covariates_matrix}{covariates matrix}

\item{weights}{perturbation weights}

\item{additional_matrix}{additional matrix}

\item{gamma}{highest polynomial order, default is 10}
}
\value{
optimal polynomial order
}
\description{
Polynomial basis with order minimizing generalized BIC
}
