% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Data generation.}
\usage{
DataGeneration(
  n_labeled,
  N_unlabeled,
  prot_att_prevalence,
  model = "correct",
  b0 = NULL,
  p = 10,
  rho = 0.4,
  n_class = 2
)
}
\arguments{
\item{n_labeled}{Number of labeled examples.}

\item{N_unlabeled}{Number of unlabeled examples.}

\item{prot_att_prevalence}{Prevalence of the protected attribute
If the number of class is more than two, the length of this vector
should be equal to the number of class. If the number of class is two,
the length of this vector can either be one or two.}

\item{model}{Indicator to generate from which model.
Default is logistic model.
possible options: "correct", "incorrect outcome", "incorrect outcome and imputation 1", and "incorrect outcome and imputation 2".}

\item{b0}{Vector of parameters to generate Y.}

\item{p}{Number of covariates. Default is 10.}

\item{rho}{Correlation between covariates. Default is 0.2.}

\item{n_class}{Number of class in the protected attribute. Default is two}

\item{Y_prevalence}{Prevalence of Y.}

\item{mis_para}{Vector of parameters to generate S when model is "mis".}
}
\value{
Data.frame.
}
\description{
Data generation.
}
