% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{DataGeneration}
\alias{DataGeneration}
\title{Data generation.}
\usage{
DataGeneration(n_labeled, N_unlabeled, prot_att_prevalence, model, rho = 0.4)
}
\arguments{
\item{n_labeled}{Number of labeled examples.}

\item{N_unlabeled}{Number of unlabeled examples.}

\item{prot_att_prevalence}{Prevalence of the protected attribute
If the number of class is more than two, the length of this vector
should be equal to the number of class. If the number of class is two,
the length of this vector can either be one or two.}

\item{model}{Indicator to generate from which model.
possible options: "scenario 1", "scenario 2"}

\item{rho}{Correlation between covariates. Default is 0.4.}
}
\value{
Data.frame.
}
\description{
Data generation.
}
