% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyBasis.R
\name{find_alpha_glm}
\alias{find_alpha_glm}
\title{Polynomial basis with order minimizing generalized BIC}
\usage{
find_alpha_glm(
  Y,
  covariates_matrix,
  additional_matrix = NULL,
  gamma = 10,
  weights = NULL,
  k = 10
)
}
\arguments{
\item{Y}{binary response variable}

\item{covariates_matrix}{covariates matrix}

\item{additional_matrix}{additional matrix}

\item{gamma}{highest polynomial order, default is 10}

\item{weights}{perturbation weights}
}
\value{
optimal polynomial order
}
\description{
Polynomial basis with order minimizing generalized BIC
}
