% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputationQuality.R
\name{ImputeQuality}
\alias{ImputeQuality}
\title{Imputation Quality}
\usage{
ImputeQuality(
  Y,
  S,
  A,
  threshold = 0.5,
  X = NULL,
  basis = "Poly(S)",
  k = 10,
  alphas = NULL,
  ...
)
}
\description{
Imputation Quality
}
