% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{RidgeRegression}
\alias{RidgeRegression}
\title{Ridge regression.}
\usage{
RidgeRegression(
  X,
  y,
  weights = NULL,
  exponents = seq(1.05, 2.5, length.out = 100)
)
}
\arguments{
\item{X}{Covariate matrix.}

\item{y}{Numeric outcome vector.}

\item{weights}{Numeric vector of weights.}

\item{exponents}{Sequence of exponents for generating lambda values.}

\item{coef}{Coefficient for scaling lambda values.}

\item{family}{Exponential family of interest.}
}
\value{
Vector containing regression coefficients.
}
\description{
Ridge regression.
}
