% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedFairness.R
\name{SSFairness}
\alias{SSFairness}
\title{Semi-supervised fairness estimation.}
\usage{
SSFairness(
  Y,
  S,
  A,
  threshold = 0.5,
  X = NULL,
  basis = "Poly(S)",
  nknots = 3,
  W = NULL
)
}
\arguments{
\item{Y}{Outcome, can contain missing values.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Threshold for classification based on the model score.
Default value is 0.5.}

\item{X}{Optional covariates matrix to be adjusted in the semi-supervised setting.
Default is NULL.}

\item{basis}{Different ways to construct W. Default is "ridge(S + X)". Other
options are "Poly(S)", "Poly(S) + X", "Spline(S)", "Spline(S) + X", "glm(S)",
"glm(S + X)", "ridge(S)", "ridge(S + X)"}

\item{nknots}{Number of knots (only used when basis = "Spline(S)" or
"Spline(S) + X"). Default is 3. Ignored otherwise.}
}
\value{
List of estimated fairness metrics and their variances.
}
\description{
Semi-supervised fairness estimation.
}
