% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemiSupervisedFairness.R
\name{SSFairness}
\alias{SSFairness}
\title{Semi-supervised fairness estimation.}
\usage{
SSFairness(Y, S, A, threshold = 0.5, W = NULL, basis = "Poly(S)")
}
\arguments{
\item{Y}{Outcome, can contain missing values.}

\item{S}{Model score.}

\item{A}{Group indicator.}

\item{threshold}{Threshold for classification based on the model score.
Default value is 0.5.}

\item{W}{Optional covariates matrix to be adjusted in the semi-supervised setting.
Default is NULL.}

\item{basis}{Different ways to construct W. Options are "Poly(S)", "Poly(S, W)"}
}
\value{
List of estimated fairness metrics and their variances.
}
\description{
Semi-supervised fairness estimation.
}
