% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupervisedFairness.R
\name{SupervisedFairness}
\alias{SupervisedFairness}
\title{Supervised fairness estimation.}
\usage{
SupervisedFairness(Y, S, A, threshold = 0.5)
}
\arguments{
\item{Y}{Outcome in the observed dataset.}

\item{S}{Model score in observed dataset.}

\item{A}{Group indicator in observed  dataset.}

\item{threshold}{Threshold for classification based on the model score.
Default value is 0.5.}
}
\description{
Supervised fairness estimation.
}
